+++ CHANGELOG.md
@@ -7,14 +7,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -->
 
-## [Development]
+## [0.4.0] - 2025-02-19
 
 ### New features
 
+* Added a `ChatVertex()` class to interact with Google Cloud's Vertex AI. (#50)
 * Added `.app(*, echo=)` support. This allows for chatlas to change the echo behavior when running the Shiny app. (#31)
+
+### Improvements
+
+* Migrated `ChatGoogle()`'s underlying python SDK from `google-generative` to `google-genai`. As a result, streaming tools are now working properly. (#50)
 
 ### Bug fixes
 
+* Fixed a bug where synchronous chat tools would not work properly when used in a `_async()` context. (#56)
 * Fix broken `Chat`'s Shiny app when `.app(*, stream=True)` by using async chat tools. (#31)
 * Update formatting of exported markdown to use `repr()` instead of `str()` when exporting tool call results. (#30)